Move check_ext_var to utility module
